MIAPIJWT/requirements.txt
#-----INSTALAR LAS DEPENDENCIAS PARA TRABAJAR-------

fastapi
uvicorn[standard]

python-jose[cryptography]
passlib[bcrypt]
python-multipart